classification/requirements.txt
# Classification Module Dependencies

# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0

# For vector-based RAG
sentence-transformers>=2.2.0
scipy>=1.10.0
numpy>=1.24.0
